requirements.txt
langchain
langchain-ollama
python-dotenv
requests
rich
ollama
